obd: don't try to handle unsupported modes
If we try to handle them, we will probably see crashes or strange behavior, as
some of them have different parsing formats. At the least, we'll see
unrecognized PIDs. Probably safer to just not-handle them upfront so that the
code can be cleanly added all at once when we need it.